+++ js/ui/react/App.tsx
@@ -3,6 +3,7 @@ import { StructureBrowser } from './components/StructureBrowser.js'
 import { SearchBar } from './components/SearchBar.js'
 import { InfoPanel } from './components/InfoPanel.js'
 import { MultiSelectPanel } from './components/MultiSelectPanel.js'
+import { Toolbar } from './components/Toolbar.js'
 
 export function App() {
   return (
@@ -11,6 +12,7 @@ export function App() {
       <StructureBrowser />
       <InfoPanel />
       <MultiSelectPanel />
+      <Toolbar />
     </>
   )
 }